Redux example
This is done in Javascript. It's based on react-redux but also we have redux-saga in place. All network requests are done via sagas as side effects of redux actions. A lot of firebase data manipulation is also done via sagas since there is no backend in behind.

This is a very good example of redux usage. Take a look to sagas.js file where a majority of logic is written.